Assistant checkpoint: Fix category creation validation and type handling
Assistant generated file changes:
- shared/schema.ts: Update category validation schema
- client/src/pages/admin/dashboard.tsx: Fix category creation data types
- server/routes.ts: Improve category creation error handling

---

User prompt:

im still getting this error when trying to create a category. the category name should be validated and stored as a string, and the default price should be validated and stored as an integer. whats going wrong?

+++ shared/schema.ts
@@ -90,7 +90,7 @@ export const cartItemsRelations = relations(cartItems, ({ one }) => ({
 export const insertCategorySchema = createInsertSchema(categories)
   .pick({
     name: true,
-    defaultPrice: z.number().min(0),
+    defaultPrice: z.number().int().min(0),
   });
 
 export const insertUserSchema = createInsertSchema(users)
